backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-socketio>=5.3.0
google-genai>=1.0.0
python-dotenv==1.0.0
google-cloud-speech==2.28.0
google-cloud-texttospeech==2.19.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
pymongo>=4.0.0
cryptography>=41.0.0
celery[redis]>=5.3.0
redis>=5.0.0
websockets>=12.0

